+++ python/cuml/cuml/datasets/blobs.py
@@ -21,10 +21,8 @@ import cupy as cp
 import numpy as np
 
 import cuml.internals
+import cuml.internals.nvtx as nvtx
 from cuml.datasets.utils import _create_rs_generator
-from cuml.internals.safe_imports import gpu_only_import_from, null_decorator
-
-nvtx_annotate = gpu_only_import_from("nvtx", "annotate", alt=null_decorator)
 
 
 def _get_centers(rs, centers, center_box, n_samples, n_features, dtype):
@@ -83,7 +81,7 @@ def _get_centers(rs, centers, center_box, n_samples, n_features, dtype):
     return centers, n_centers
 
 
-@nvtx_annotate(message="datasets.make_blobs", domain="cuml_python")
+@nvtx.annotate(message="datasets.make_blobs", domain="cuml_python")
 @cuml.internals.api_return_generic()
 def make_blobs(
     n_samples=100,
