+++ python/cuml/cuml/datasets/regression.pyx
@@ -23,14 +23,13 @@ import numpy as np
 from pylibraft.common.handle import Handle
 
 import cuml.internals
+import cuml.internals.nvtx as nvtx
 from cuml.internals.array import CumlArray
-from cuml.internals.safe_imports import gpu_only_import_from, null_decorator
 
 from libc.stdint cimport uint64_t, uintptr_t
 from libcpp cimport bool
 from pylibraft.common.handle cimport handle_t
 
-nvtx_annotate = gpu_only_import_from("nvtx", "annotate", alt=null_decorator)
 
 cdef extern from "cuml/datasets/make_regression.hpp" namespace "ML":
     void cpp_make_regression "ML::Datasets::make_regression" (
@@ -74,7 +73,7 @@ inp_to_dtype = {
 }
 
 
-@nvtx_annotate(message="datasets.make_regression", domain="cuml_python")
+@nvtx.annotate(message="datasets.make_regression", domain="cuml_python")
 @cuml.internals.api_return_generic()
 def make_regression(
     n_samples=100,
