+++ python/cuml/cuml/ensemble/randomforestregressor.pyx
@@ -22,14 +22,13 @@ from cuml.internals.safe_imports import (
     cpu_only_import,
     gpu_only_import,
     gpu_only_import_from,
-    null_decorator,
 )
 
 np = cpu_only_import('numpy')
-nvtx_annotate = gpu_only_import_from("nvtx", "annotate", alt=null_decorator)
 rmm = gpu_only_import('rmm')
 
 import cuml.internals
+import cuml.internals.nvtx as nvtx
 from cuml.internals import logger
 from cuml.internals.array import CumlArray
 from cuml.internals.mixins import RegressorMixin
@@ -50,8 +49,6 @@ from cuml.legacy.fil.fil import TreeliteModel
 
 from libc.stdint cimport uint64_t, uintptr_t
 from libcpp cimport bool
-
-from cuml.internals.safe_imports import gpu_only_import_from
 
 cuda = gpu_only_import_from('numba', 'cuda')
 
@@ -440,7 +437,7 @@ class RandomForestRegressor(BaseRandomForestModel,
                                  algo=algo,
                                  fil_sparse_format=fil_sparse_format)
 
-    @nvtx_annotate(
+    @nvtx.annotate(
         message="fit RF-Regressor @randomforestregressor.pyx",
         domain="cuml_python")
     @generate_docstring()
@@ -563,7 +560,7 @@ class RandomForestRegressor(BaseRandomForestModel,
         del X_m
         return preds
 
-    @nvtx_annotate(
+    @nvtx.annotate(
         message="predict RF-Regressor @randomforestclassifier.pyx",
         domain="cuml_python")
     @insert_into_docstring(parameters=[('dense', '(n_samples, n_features)')],
@@ -624,7 +621,7 @@ class RandomForestRegressor(BaseRandomForestModel,
 
         return preds
 
-    @nvtx_annotate(
+    @nvtx.annotate(
         message="score RF-Regressor @randomforestclassifier.pyx",
         domain="cuml_python")
     @insert_into_docstring(parameters=[('dense', '(n_samples, n_features)'),
