+++ python/cuml/cuml/tsa/arima.pyx
@@ -16,14 +16,10 @@
 
 # distutils: language = c++
 
-from cuml.internals.safe_imports import (
-    cpu_only_import,
-    gpu_only_import_from,
-    null_decorator,
-)
+import cuml.internals.nvtx as nvtx
+from cuml.internals.safe_imports import cpu_only_import
 
 np = cpu_only_import('numpy')
-nvtx_annotate = gpu_only_import_from("nvtx", "annotate", alt=null_decorator)
 
 from libc.stdint cimport uintptr_t
 from libcpp cimport bool
@@ -462,7 +458,7 @@ class ARIMA(Base):
             return "ARIMA({},{},{}) ({}) - {} series".format(
                 order.p, order.d, order.q, intercept_str, self.batch_size)
 
-    @nvtx_annotate(message="tsa.arima.ARIMA._ic", domain="cuml_python")
+    @nvtx.annotate(message="tsa.arima.ARIMA._ic", domain="cuml_python")
     @cuml.internals.api_base_return_any_skipall
     def _ic(self, ic_type: str):
         """Wrapper around C++ information_criterion
@@ -755,7 +751,7 @@ class ARIMA(Base):
                     d_lower,
                     d_upper)
 
-    @nvtx_annotate(message="tsa.arima.ARIMA.forecast", domain="cuml_python")
+    @nvtx.annotate(message="tsa.arima.ARIMA.forecast", domain="cuml_python")
     @cuml.internals.api_base_return_generic_skipall
     def forecast(
         self,
@@ -827,7 +823,7 @@ class ARIMA(Base):
         if not hasattr(self, "sigma2_"):
             self.sigma2_ = CumlArray.empty(self.batch_size, np.float64)
 
-    @nvtx_annotate(message="tsa.arima.ARIMA._estimate_x0",
+    @nvtx.annotate(message="tsa.arima.ARIMA._estimate_x0",
                    domain="cuml_python")
     @cuml.internals.api_base_return_any_skipall
     def _estimate_x0(self):
@@ -950,7 +946,7 @@ class ARIMA(Base):
         self.unpack(self._batched_transform(x), convert_dtype)
         return self
 
-    @nvtx_annotate(message="tsa.arima.ARIMA._loglike", domain="cuml_python")
+    @nvtx.annotate(message="tsa.arima.ARIMA._loglike", domain="cuml_python")
     @cuml.internals.api_base_return_any_skipall
     def _loglike(self, x, trans=True, method="ml", truncate=0, convert_dtype=True):
         """Compute the batched log-likelihood for the given parameters.
@@ -1017,7 +1013,7 @@ class ARIMA(Base):
 
         return np.array(vec_loglike, dtype=np.float64)
 
-    @nvtx_annotate(message="tsa.arima.ARIMA._loglike_grad",
+    @nvtx.annotate(message="tsa.arima.ARIMA._loglike_grad",
                    domain="cuml_python")
     @cuml.internals.api_base_return_any_skipall
     def _loglike_grad(self, x, h=1e-8, trans=True, method="ml", truncate=0,
@@ -1140,7 +1136,7 @@ class ARIMA(Base):
 
         return np.array(vec_loglike, dtype=np.float64)
 
-    @nvtx_annotate(message="tsa.arima.ARIMA.unpack", domain="cuml_python")
+    @nvtx.annotate(message="tsa.arima.ARIMA.unpack", domain="cuml_python")
     def unpack(self, x: Union[list, np.ndarray], convert_dtype=True):
         """Unpack linearized parameter vector `x` into the separate
         parameter arrays of the model
@@ -1169,7 +1165,7 @@ class ARIMA(Base):
         cpp_unpack(handle_[0], cpp_params, order, <int> self.batch_size,
                    <double*>d_x_ptr)
 
-    @nvtx_annotate(message="tsa.arima.ARIMA.pack", domain="cuml_python")
+    @nvtx.annotate(message="tsa.arima.ARIMA.pack", domain="cuml_python")
     def pack(self) -> np.ndarray:
         """Pack parameters of the model into a linearized vector `x`
 
@@ -1193,7 +1189,7 @@ class ARIMA(Base):
 
         return d_x_array.to_output("numpy")
 
-    @nvtx_annotate(message="tsa.arima.ARIMA._batched_transform",
+    @nvtx.annotate(message="tsa.arima.ARIMA._batched_transform",
                    domain="cuml_python")
     @cuml.internals.api_base_return_any_skipall
     def _batched_transform(self, x, isInv=False):
